requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
altair==4.2.2
async-timeout==4.0.2
attrs==22.2.0
beautifulsoup4==4.11.2
blinker==1.5
blobfile==2.0.1
boto3==1.26.84
botocore==1.29.84
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
dataclasses-json==0.5.7
decorator==5.1.1
dnspython==2.3.0
entrypoints==0.4
filelock==3.9.0
Flask==2.2.3
frozenlist==1.3.3
gitdb==4.0.10
GitPython==3.1.31
googleapis-common-protos==1.56.4
grpc-gateway-protoc-gen-openapiv2==0.1.0
grpcio==1.51.3
idna==3.4
importlib-metadata==6.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.2.0
jsonschema==4.17.3
loguru==0.6.0
lxml==4.9.2
lz4==4.3.2
Markdown==3.4.1
markdown-it-py==2.2.0
markdown2==2.4.8
MarkupSafe==2.1.2
marshmallow==3.19.0
marshmallow-enum==1.5.1
mdurl==0.1.2
multidict==6.0.4
mypy-extensions==1.0.0
nltk==3.8.1
numpy==1.24.2
openai==0.27.0
packaging==23.0
pandas==1.5.3
Pillow==9.4.0
pinecone-client==2.2.1
protobuf==3.19.3
pyarrow==11.0.0
pycryptodomex==3.17
pydeck==0.8.0
Pygments==2.14.0
Pympler==1.0.1
pyrsistent==0.19.3
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
regex==2022.10.31
requests==2.28.2
requests-toolbelt==0.10.1
rich==13.3.2
s3transfer==0.6.0
semver==2.13.0
six==1.16.0
slack-sdk==3.20.0
smmap==5.0.0
soupsieve==2.4
streamlit==1.19.0
tiktoken==0.3.0
toml==0.10.2
toolz==0.12.0
tornado==6.2
tqdm==4.65.0
typing-inspect==0.8.0
typing_extensions==4.5.0
tzdata==2022.7
tzlocal==4.2
urllib3==1.26.14
validators==0.20.0
waitress==2.1.2
Werkzeug==2.2.3
yarl==1.8.2
zipp==3.15.0
